// packages/datadog-api-client-v2/models/SecurityFilterExclusionFilterResponse.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* A single exclusion filter.
*/
export declare class SecurityFilterExclusionFilterResponse {
    /**
    * The exclusion filter name.
    */
    'name'?: string;
    /**
    * The exclusion filter query.
    */
    'query'?: string;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): SecurityFilterExclusionFilterResponse;
    static serialize(data: SecurityFilterExclusionFilterResponse): {
        [key: string]: any;
    };
    constructor();
}